reverted DRIVER_INTERFACE type to uint16_t

+++ libindi/libs/indibase/basedevice.cpp
@@ -1394,7 +1394,7 @@ const char *INDI::BaseDevice::getDriverVersion()
     return nullptr;
 }
 
-uint32_t INDI::BaseDevice::getDriverInterface()
+uint16_t INDI::BaseDevice::getDriverInterface()
 {
     ITextVectorProperty *driverInfo = getText("DRIVER_INFO");
